refactor(cmd/reverst): move listening and serving into internal/server type
